Split `job_logger` into `mod.rs` and `logger.rs`.
